+++ src/app/modules/recipes/list/list.component.ts
@@ -1,4 +1,6 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input, OnInit, SimpleChange } from '@angular/core';
+import { Router } from '@angular/router';
+import { DialogHelperService } from 'src/app/shared/services/dialog-helper.service';
 import { Recipe } from '../models/recipe.model';
 import { RecipeService } from '../services/recipe.service';
 
@@ -13,7 +15,7 @@ export class ListComponent implements OnInit {
   @Input()
   recipesList: Recipe[] = []
 
-  constructor(private recipeService: RecipeService, private changeDetectorRef: ChangeDetectorRef) {
+  constructor(private recipeService: RecipeService, private changeDetectorRef: ChangeDetectorRef, private dialogHelperService: DialogHelperService, private router: Router) {
 
   }
 
@@ -22,8 +24,18 @@ export class ListComponent implements OnInit {
   }
 
   deleteOneRecipe(id: string) {
-    this.recipeService.deleteRecipe(id).subscribe(
-      () => this.recipeService.updateRecipesList()
+    this.dialogHelperService.openDialogYesNo('Warning', 'Are you sure to delete this recipe?').subscribe(
+      dialogResult => {
+        if (dialogResult === true) {
+          this.recipeService.deleteRecipe(id).subscribe(
+            () => {
+              this.recipeService.updateRecipesList();
+              this.router.navigate(['addRecipe']);
+            }
+          );
+        }
+      }
     )
+
   }
 }
